backend/compfest-ai/requirements.txt
flask
flask-cors
python-dotenv
marshmallow
